### 2024-09-07 #7190 @editorialbot
[PRE REVIEW]: kugelpy: A Python package for modeling pebble bed reactor run-in
**Submitting author:** @ethan-fowler (<a href="http://orcid.org/0009-0005-5922-8389">Ethan Fowler</a>)
**Repository:** https://github.com/idaholab/kugelpy.git
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.1.0
