### 2024-12-11 #7190 @kellyrowland
hello @jpmorgan98 @giozu 👋 would you be interested in and available for reviewing this JOSS submission?
