### 2025-01-13 #7190 @kellyrowland
hello @gwenchee @andrewryh 👋 would you be interested in and available for reviewing this JOSS submission?
