### 2025-01-23 #7190 @kellyrowland
hello @sskutnik @damar-wicaksono 👋 would you be interested in and available for reviewing this JOSS submission?
