### 2025-03-12 #7727 @damar-wicaksono
@kellyrowland: I am currently reviewing the submission, and while I have some remarks and suggestions, one key aspect remains unclear to me: evaluating the functionality of the package (i.e., "Functionality: Have the functional claims of the software been confirmed?").

As I understand it, kugelpy relies on Serpent code (perhaps I should have realized this much earlier)—some of the package's main functionalities include building the input deck, executing the code, and parsing the relevant output files. However, since I no longer have access to Serpent, I’m uncertain whether I can finalize the review.

Do you have any suggestions for how I should proceed, aside from recusing myself from the review?
