### 2025-03-12 #7727 @kellyrowland
@damar-wicaksono thanks for flagging this up - I think we'll want to call on another reviewer who's able to verify the functionality by way of checking things in the context of using them with Serpent. I can try to reach our to additional reviewers - any chance that you might be able to recommend someone?

@ethan-fowler happy to open an issue over on the code repo if you'd like, but I just caught that the [Specific guides](https://github.com/idaholab/kugelpy?tab=readme-ov-file#specific-guides) section of the README doesn't link to any guides (the link in the first item is a 404, and the second item doesn't have a linkyet)
