### 2025-03-26 #7727 @kellyrowland
thanks! @ZoeRichter would you be interested in and available to review this JOSS submission? you'll need to have active access to Serpent to test the code out.
