### 2025-03-27 #7727 @ZoeRichter
Hello - yes, I'd be interested and available to review this submission, and I have current access to Serpent.  This would be my first review with JOSS - is there anything special I need to do before getting started, or is the reviewer guide and checklist a good place to start?
