### 2025-04-04 idaholab/kugelpy#4 @sskutnik
JOSS review: Documentation defects
There are a few broken links throughout the main README.md document. For example, it appears the naming convention was changed, wherein several links point to `kugelpy/docs/pbr_runin.md` instead of `kugelpy/docs/run_in.md`

It would be extremely useful to link users to what appears to be the [top-level documentation page](https://github.com/idaholab/kugelpy/blob/main/docs/kugelpy.md)

The ["General Overview"](https://github.com/idaholab/kugelpy/blob/main/docs/kugelpy.md) section of the main document could use some attention for formatting; mostly it seems like the intended markdown didn't carry over correctly to GitHub. Direct links to the subordinate modules here would be useful.
