### 2025-04-04 #7727 @sskutnik
@kellyrowland @ethan-fowler I've opened an issue on some of the [documentation issues](https://github.com/idaholab/kugelpy/issues/4); mostly just a matter of getting the right links in place. For instance, what I believe to be the [main documentation page](https://github.com/idaholab/kugelpy/blob/main/docs/kugelpy.md) has a point of contact for support and links to relevant examples but nothing's linking to it.

There's not really a formal "contributing" guideline but I'm not going to be terribly picky over that; mostly I think having a point of contact would be a good starting point.

PS: Sorry it's taken me so long to get rolling on this review.
