### 2025-04-04 idaholab/kugelpy#5 @sskutnik
JOSS review: Failing unit tests
On my local install, when I run the unit test suite as-indicated in the instructions (i.e., `py.test -vv -./kugelpy/tests`), I get five failing tests (out of 56):

```
FAILED kugelpy/tests/test_pebble_bed_reactor.py::test_build_all_blocks
FAILED kugelpy/tests/test_pebble_bed_reactor.py::test_build_pbr_core
FAILED kugelpy/tests/test_pebble_bed_reactor.py::test_build_pbr_core_no_dimples
FAILED kugelpy/tests/test_pebble_bed_reactor.py::test_build_pbr_core_simple
FAILED kugelpy/tests/test_pebble_sorter.py::test_read_burn_material
```
