### 2025-04-04 #7727 @sskutnik
@kellyrowland @ethan-fowler Tests are present but when I run locally I'm getting five test failures; [opened a corresponding issue](https://github.com/idaholab/kugelpy/issues/5).
