### 2025-04-04 idaholab/kugelpy#6 @sskutnik
JOSS review: Examples directory is missing
On the [main README](https://github.com/idaholab/kugelpy?tab=readme-ov-file#main-folder-files), there is a reference to a top-level folder "examples" which does not appear to exist. (Possibly did not get ported over in the move to Github.)
