### 2025-04-04 #7727 @sskutnik
@kellyrowland @ethan-fowler Examples directory is referenced in the documentation but [https://github.com/idaholab/kugelpy/issues/6](doesn't seem to be on the main Github page).
