### 2025-04-09 #7727 @ZoeRichter
I wanted to give an update - I'm having some unrelated hardware issues on my end, and while my desktop is (mostly) working, it's getting a bit fussy.  Worst case scenario, I think I can review on a fresh install on new hardware in a few weeks, once the new desktops are ordered and come in.  Also - could you confirm the minimum Serpent version that kugelpy can work with?

In the meantime, I can at least get started on reviewing what I can.
