### 2025-04-09 idaholab/kugelpy#7 @ZoeRichter
JOSS Review: Author list discrepancy
To satisfy:  "Contribution and authorship: Has the submitting author (@ethan-fowler) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?"  I looked at the contributions and commit/PR history for this repo - but it looks like this was ported over from another project, and it seems like much of the information about who contributed what may have been lost.

The paper's author list is Ethan Fowler, Ryan Stewart, Paolo Balestra, and Jack Cavaluzzi - since the history isn't here, can you confirm or briefly describe their code contributions for the sake of review?  The commits only show Ryan merging PRs with commits from Ethan.

Regardless of the above - setup.py only lists Ryan Stewart and Paolo Balestra as authors, so that should be updated (you can probably also remove the commented out dummy email line, if you aren't providing those).
