### 2025-04-09 #7727 @ZoeRichter
I'm going through a few quick things I can check, and I noticed [an issue with the author list](https://github.com/idaholab/kugelpy/issues/7).

The quick version is that the setup.py doesn't list all four authors, and because this project was ported over from outside github, I'd like to double check code contributions from the authors with you before I check it off, since I suspect a lot of the code's history was lost.
