### 2025-04-09 idaholab/kugelpy#8 @ZoeRichter
JOSS Review: Paper figure comments
- Figure too small
- Axial cross section on the right has an incorrect aspect ratio, which is distorting the geometry.  I would recommend either splitting the left and right views into two images, or basing the aspect ratio of the combined figure off the image on the right, and adding white space to the top and bottom of the left hand view to compensate.
- I can tell what the arrows pointing to the radial view are trying to show, but it is unclear what the arrows are pointing to or trying to convey on the axial side.
- I think it would be useful to add a legend explaining what materials the colors correspond to.
