### 2025-04-09 #7727 @ZoeRichter
I have made an issue for [my comments on the paper's figure.](https://github.com/idaholab/kugelpy/issues/8)
