### 2025-04-10 idaholab/kugelpy#7 @ZoeRichter
I'd also like to clarify that I'm not expecting a super detailed list of who exactly wrote what - something brief ("this person worked on xyz part of the code/paper/etc") is fine.
