### 2025-04-14 idaholab/kugelpy#9 @damar-wicaksono
JOSS Review (Reviewer #2): Software Paper
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/7727)

Please find below some questions, remarks, and suggestions regarding the draft paper. Please feel free to let me know if you need clarification. Thanks!

## Summary

- I think the summary can be simplified moving much of the early sentences to the Statement of Need. Specifically, describing pebble bed reactors in such details may not be necessary here.
- I would then suggest to focus on the particular problems that `kugelpy` try to address (starting, I believe, l. 14) and what `kugelpy` actually does to address those problems.
- I think, stating the components of `kugelpy`, i.e., `sea_serpent`, `kugelpy`, `first_mate` (last sentence) are too low-level details to be in the summary. On the other hand, it is also not very clear to general audience "general Serpent reactor functions" or "basic utilities" actually are. Perhaps for the Summary section, simply provide a concrete list of what `kugelpy` does.
- l. 7 "...that make use of passive safety features which operate **independent** of human input": independently (?).
- ll. 9-10 "...graphite pebbles several centimeters in radius **containing on the order of a few grams of nuclear fuel each.**": Perhaps for slightly clearer version - "..., each containing approximately a few grams of nuclear fuel."

## Statement of need

- As suggested above, the introduction to PBRs and the difficulties of their modeling can be succinctly described here before moving on to the details of `kugelpy`'s capabilities.
- I think, this section can be split into multiple paragraphs instead of one long paragraph.
- In relation to the point above, I would suggest that when describing each of the `kugelpy`'s capabilities, it is preceded with the actual problems of modeling PBRs to have a better context.
- I would also suggest to avoid referring to a function's name or object's name (as they may change in the future as unlikely as they may be right now) but focus on their use in the context of PBRs modeling.
- In relation to the point above, it is not clear what `PebbleSorter` object actually means, so perhaps first provide clear description on what kind of objects that `kugelpy` are actually modeling.
- ll. 27 - 29 "`kugelpy` is a Python module that is capable of producing Serpent input files for high-temperature gas cooled PBRs, **similar to the AVR, HTR-10, or Xe-100 (Mulder & Boyes, 2020; Pohl, 2006; Wu et al., 2002).**": "similar to" here might be misleading if this list refers to other Python modules (similar to `kugelpy`) instead of other reactors (similar to PBRs).

## State of the field

Is there any other available open-source tools that are similar to `kugelpy` or offer similar functionalities? If not, perhaps this could be clearly stated in the Statement of need section.

## Example

- Similar to my previous comment, I would avoid mentioning particular filenames for the examples in the paper, e.g., `run_in.md`, `raw_dist.inp,` (as the may change in the future; for instance, I cannot find `raw_dist.inp`); perhaps focus on describing what examples are available, their reference design (i.e, GPBR200), and what the users can do to adapt to their needs.
- `run_in.md` is a markdown file, I don't think they qualify as a (annotated) python script. The file actually refers to itself as a "notebook".
- l. 61 "...variety of means including regular lattices, DEM simulations, or random distributions.": perhaps expand DEM as it is the first (and only) time it appears in the paper.
