### 2025-04-14 #7727 @damar-wicaksono
As I previously mentioned to @kellyrowland,  I won't be able to fully review the package's functionality due to lack of access to Serpent. Nevertheless, since I had already begun the review, I opened two issues in the repository with my comments and suggestions regarding the [software paper](https://github.com/idaholab/kugelpy/issues/9) and [documentation (and testing)](https://github.com/idaholab/kugelpy/issues/10). Perhaps they can be useful.

I would also second the issue raised by @ZoeRichter regarding [the author list](https://github.com/idaholab/kugelpy/issues/7).
